Move BiometricActivity to modal package
