+++ app/src/main/AndroidManifest.xml
@@ -42,7 +42,7 @@
             android:theme="@style/AppTheme.NoActionBar" />
         <activity android:name="nl.tudelft.cs4160.identitychain.chainExplorer.ChainExplorerActivity" />
         <activity
-            android:name="nl.tudelft.cs4160.identitychain.main.BiometricActivity"
+            android:name="nl.tudelft.cs4160.identitychain.modals.BiometricActivity"
             android:theme="@style/AppTheme.NoActionBar"/>
     </application>
